Move smart order from dialog to dedicated full page with delivery days
- Create /admin/orders/create page with full-width layout (8/4 grid split)
  for product selection and cart sidebar instead of cramped modal
- Auto-calculate expected delivery date per supplier using leadTimeDays
  field (already in schema) — shown in cart per supplier group
- Add leadTimeDays to /api/suppliers/products response
- Remove ~640 lines of dialog code from orders page, replaced with
  a simple link to the new page
